# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz API",
        "description": "API Documentation",
        "contact": {
            "name": "API Support",
            "email": "marko.mandic.engr@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/variants/{variant_id}/images": {
            "get": {
                "tags": [
                    "Variant Images"
                ],
                "summary": "List variant images",
                "operationId": "23c7bfe6831b66bf9ade38b4f4656264",
                "parameters": [
                    {
                        "name": "variant_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "404": {
                        "description": "Variant not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Variant Images"
                ],
                "summary": "Create a new variant image",
                "operationId": "45156ae8ceeb2386ccc175f0efacee92",
                "parameters": [
                    {
                        "name": "variant_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Variant image created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/variant-images/{id}": {
            "get": {
                "tags": [
                    "Variant Images"
                ],
                "summary": "Get a specific variant image",
                "operationId": "ccb555f130b4934c9ab785bab1f0e07f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "404": {
                        "description": "Variant image not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Variant Images"
                ],
                "summary": "Update a variant image",
                "operationId": "5eb82bc7f9f67426b93af16bfbd1917e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Variant image updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Variant image not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Variant Images"
                ],
                "summary": "Delete a variant image",
                "operationId": "12f74407eb7d1f9af23c5e5e9a8311d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Variant image deleted successfully"
                    },
                    "404": {
                        "description": "Variant image not found"
                    }
                }
            }
        }
    },
...
}